Refactor ObservableArrayChangeEvent
Replace extension with static equality function with just static
equality function for the sake of readability. Also, it's not possible
to create such extension for checking equality of two arrays that
contains ObservableArrayChangeEvent objects. It is possible using
plain equality function (not embedded as a static extension function),
so to keep it consistent, will use plain equality functions.